requirements.txt
grpcio
htmldocx
git+https://github.com/Saritasa/html_docx
lxml
Markdown
protobuf
pylint
pytest
python-docx
PyYAML
requests
yattag
matplotlib
inflection
openpyxl
neo4j
iterfzf
Jinja2
paramiko
GitPython
pycryptodome
smbus
pillow
tkterminal
packages/epsmsgbus
